webhook-cloud-function/requirements.txt
flask==2.3.1
google-cloud-firestore==2.11.1
google-cloud-secret-manager==2.7.0
openai==0.27.5
requests==2.27.1